+++ packages/pkg/src/config/index.ts
@@ -1,0 +1,2 @@
+export { default as getUserConfig } from './userConfig.js';
+export { default as getCliOptions } from './cliOptions.js';