+++ packages/pkg/src/helpers/getTaskIO.ts
@@ -1,47 +1,43 @@
-import { join } from 'path';
-import { isDirectory, isFile } from '../utils.js';
+import { resolve } from 'path';
+import { isDirectory } from '../utils.js';
 import { TaskName } from '../types.js';
 
+export function formatEntry(inputEntry: string | string[] | Record<string, string>): Record<string, string> {
+  const entry = {};
+  if (typeof inputEntry === 'string') {
+    entry[getEntryId(inputEntry)] = inputEntry;
+  } else if (Array.isArray(inputEntry)) {
+    inputEntry.forEach((item) => {
+      entry[getEntryId(item)] = item;
+    });
+  } else if (typeof inputEntry === 'object') {
+    Object.keys(inputEntry).forEach((key) => {
+      entry[key] = inputEntry[key];
+    });
+  }
+  return entry;
+}
+
+// Eg. src/index.js => index
+function getEntryId(entry: string): string {
+  return entry.split('/').pop().split('.').shift();
+}
+
 const DEFAULT_ENTRY_DIR = 'src';
-const DEFAULT_ENTRY_FILE = [
-  'index.js',
-  'index.ts',
-  'index.mts',
-  'index.mjs',
-  'index.tsx',
-  'index.jsx',
-];
 
-export const findDefaultEntryFile = (path: string) => {
-  return DEFAULT_ENTRY_FILE
-    .map((value) => join(path, value))
-    .find((file) => isFile(file));
-};
+export const getDefaultEntryDir = (rootDir: string, entryDir = DEFAULT_ENTRY_DIR) => {
+  entryDir = resolve(rootDir, entryDir);
 
-export const getEntryFile = (rootDir: string) => {
-  const defaultEntryDir = getEntryDir(rootDir);
-  const entryFile = findDefaultEntryFile(defaultEntryDir);
-
-  if (entryFile === undefined) {
-    throw new Error(`Could not find path ${entryFile}, which be regarded as entry`);
+  if (!isDirectory(entryDir)) {
+    throw new Error(`Failed to resolve ${entryDir} as an entry directory.`);
   }
 
-  return entryFile;
-};
-
-export const getEntryDir = (rootDir: string) => {
-  const defaultEntryDir = join(rootDir, DEFAULT_ENTRY_DIR);
-
-  if (isDirectory(defaultEntryDir)) {
-    return defaultEntryDir;
-  }
-
-  throw new Error(`Failed to resolve ${defaultEntryDir} as entry directory`);
+  return entryDir;
 };
 
 export const getOutputDir = (rootDir: string, taskName: TaskName) => {
   if (taskName.includes('transform')) {
-    return join(rootDir, taskName.split('-')[1]);
+    return resolve(rootDir, taskName.split('-')[1]);
   }
-  return join(rootDir, 'dist');
+  return resolve(rootDir, 'dist');
 };