+++ packages/plugin-docusaurus/src/Previewer/Mobile.tsx
@@ -1,5 +1,6 @@
-import React, { useRef } from 'react';
+import React, { useRef, useState } from 'react';
 import CodeBlock from '@theme-original/CodeBlock';
+import { QRCodeSVG } from 'qrcode.react';
 import styles from './styles.module.css';
 import './styles.css';
 
@@ -15,13 +16,28 @@ function FullScreenSvg(props) {
   );
 }
 
+function QrCodeSvg(props) {
+  return (
+    <svg width="20px" height="20px" viewBox="0 0 1024 1024" {...props} version="1.1" xmlns="http://www.w3.org/2000/svg" p-id="2552"><path d="M468 128H160c-17.7 0-32 14.3-32 32v308c0 4.4 3.6 8 8 8h332c4.4 0 8-3.6 8-8V136c0-4.4-3.6-8-8-8z m-56 284H192V192h220v220z" p-id="2553" fill="#666666"></path><path d="M274 338h56c4.4 0 8-3.6 8-8v-56c0-4.4-3.6-8-8-8h-56c-4.4 0-8 3.6-8 8v56c0 4.4 3.6 8 8 8zM468 548H136c-4.4 0-8 3.6-8 8v308c0 17.7 14.3 32 32 32h308c4.4 0 8-3.6 8-8V556c0-4.4-3.6-8-8-8z m-56 284H192V612h220v220z" p-id="2554" fill="#666666"></path><path d="M274 758h56c4.4 0 8-3.6 8-8v-56c0-4.4-3.6-8-8-8h-56c-4.4 0-8 3.6-8 8v56c0 4.4 3.6 8 8 8zM864 128H556c-4.4 0-8 3.6-8 8v332c0 4.4 3.6 8 8 8h332c4.4 0 8-3.6 8-8V160c0-17.7-14.3-32-32-32z m-32 284H612V192h220v220z" p-id="2555" fill="#666666"></path><path d="M694 338h56c4.4 0 8-3.6 8-8v-56c0-4.4-3.6-8-8-8h-56c-4.4 0-8 3.6-8 8v56c0 4.4 3.6 8 8 8zM888 548h-48c-4.4 0-8 3.6-8 8v134h-78V556c0-4.4-3.6-8-8-8H556c-4.4 0-8 3.6-8 8v332c0 4.4 3.6 8 8 8h48c4.4 0 8-3.6 8-8V644h78v102c0 4.4 3.6 8 8 8h190c4.4 0 8-3.6 8-8V556c0-4.4-3.6-8-8-8z" p-id="2556" fill="#666666"></path><path d="M746 832h-48c-4.4 0-8 3.6-8 8v48c0 4.4 3.6 8 8 8h48c4.4 0 8-3.6 8-8v-48c0-4.4-3.6-8-8-8zM888 832h-48c-4.4 0-8 3.6-8 8v48c0 4.4 3.6 8 8 8h48c4.4 0 8-3.6 8-8v-48c0-4.4-3.6-8-8-8z" p-id="2557" fill="#666666"></path></svg>
+  );
+}
+
 function MobilePreview({ code, url }) {
   const iframe = useRef(null);
+  const [showDialog, setShowDialog] = useState(false);
   const redirctFullScreen = () => {
     window.open(url);
   };
   const reloadIframe = () => {
     iframe?.current?.contentWindow.location.reload();
+  };
+
+  const openQRCode = () => {
+    setShowDialog(true);
+  };
+
+  const closeQrCode = () => {
+    setShowDialog(false);
   };
 
   return (
@@ -45,6 +61,10 @@ function MobilePreview({ code, url }) {
             className={styles.operationItem}
             onClick={redirctFullScreen}
             ><FullScreenSvg />  &nbsp;全屏模式</div>
+          <div
+            className={styles.operationItem}
+            onClick={openQRCode}
+            ><QrCodeSvg /> &nbsp;二维码</div>
         </div>
         <div className={styles.iframeWrapper}>
           <iframe
@@ -54,6 +74,9 @@ function MobilePreview({ code, url }) {
             frameBorder="0"
             src={url}
             />
+          {showDialog && <div className={styles.dialog} onClick={closeQrCode}>
+            <QRCodeSVG value={`${location.origin}${url}`} size={100}></QRCodeSVG>
+          </div>}
         </div>
 
       </div>